// resources/types/libs.ts
export type TEditBlockDispatch = {
	clearSelectedBlock: () => any
	duplicateBlocks: ( wrapperIds: string[] ) => void
	enterFormattedText: () => any
	exitFormattedText: () => any
	failResolution: () => any
	failResolutions: () => any
	finishResolution: () => any
	finishResolutions: () => any
	flashBlock: () => any
	hideInsertionPoint: () => any
	insertAfterBlock: () => any
	insertBeforeBlock: () => any
	insertBlock: ( block: TBlock, index: number, blockRootId: string, bool: boolean ) => void
	insertBlocks: () => any
	insertDefaultBlock: () => any
	invalidateResolution: () => any
	invalidateResolutionForStore: () => any
	invalidateResolutionForStoreSelector: () => any
	mergeBlocks: () => any
	moveBlockToPosition: () => any
	moveBlocksDown: () => any
	moveBlocksToPosition: ( clientId: string[], wrapperClientId: string, blockClientId: string, index: number ) => void
	moveBlocksUp: () => any
	multiSelect: () => any
	receiveBlocks: () => any
	removeBlock: () => any
	removeBlocks: () => any
	replaceBlock: () => any
	replaceBlocks: () => any
	replaceInnerBlocks: ( clientId: string, innerBlocks: object[], bool?: boolean ) => any
	resetBlocks: () => any
	resetSelection: () => any
	selectBlock: () => any
	selectNextBlock: () => any
	selectPreviousBlock: () => any
	selectionChange: () => any
	setBlockMovingClientId: () => any
	setBlockVisibility: () => any
	setHasControlledInnerBlocks: () => any
	setNavigationMode: () => any
	setTemplateValidity: () => any
	showInsertionPoint: () => any
	startDraggingBlocks: () => any
	startMultiSelect: () => any
	startResolution: () => any
	startResolutions: () => any
	startTyping: () => any
	stopDraggingBlocks: () => any
	stopMultiSelect: () => any
	stopTyping: () => any
	synchronizeTemplate: () => any
	toggleBlockHighlight: () => any
	toggleBlockMode: () => any
	toggleSelection: () => any
	updateBlock: () => any
	updateBlockAttributes: ( clientId: string, attributes: object ) => any
	updateBlockListSettings: () => any
	updateSettings: () => any
	validateBlocksToTemplate: () => any
	__unstableDeleteSelection: () => any
	__unstableExpandSelection: () => any
	__unstableMarkAutomaticChange: () => any
	__unstableMarkLastChangeAsPersistent: () => any
	__unstableMarkNextChangeAsNotPersistent: () => any
	__unstableSaveReusableBlock: () => any
	__unstableSetEditorMode: () => any
	__unstableSetTemporarilyEditingAsBlocks: () => any
	__unstableSplitSelection: () => any
}

export type TBlock = {
	name: string
	clientId: string
	attributes: {
		[key: string]: any
	}
	innerBlocks: object[]
	isValid: boolean
}

export type TRegistry = {
	batch: () => any
	dispatch: () => any
	namespaces: object
	register: () => any
	registerGenericStore: () => any
	registerStore: () => any
	resolveSelect: () => any
	select: ( store: { instantiate: () => any, name: string } ) => TSelectFunctions
	stores: object
	subscribe: () => any
	suspendSelect: () => any
	use: () => any
	__unstableMarkListeningStores: () => any
	__unstableSubscribeStore: () => any
}

export type TSelectFunctions = {
	areInnerBlocksControlled: () => any
	canEditBlock: () => any
	canInsertBlockType: () => any
	canInsertBlocks: () => any
	canLockBlockType: () => any
	canMoveBlock: () => any
	canMoveBlocks: () => any
	canRemoveBlock: () => any
	canRemoveBlocks: () => any
	didAutomaticChange: () => any
	getAdjacentBlockClientId: () => any
	getBlock: ( wrapperClientId?: string ) => TBlock[]
	getBlockAttributes: () => any
	getBlockCount: () => any
	getBlockHierarchyRootClientId: () => any
	getBlockIndex: ( wrapperClientId?: string ) => number
	getBlockInsertionPoint: () => any
	getBlockListSettings: () => any
	getBlockMode: () => any
	getBlockName: ( wrapperClientId: string ) => any
	getBlockOrder: ( clientId: string ) => any
	getBlockParents: () => any
	getBlockParentsByBlockName: () => any
	getBlockRootClientId: ( id: string ) => any
	getBlockSelectionEnd: () => any
	getBlockSelectionStart: () => any
	getBlockTransformItems: () => any
	getBlocks: ( clientId: string ) => any
	getBlocksByClientId: () => any
	getCachedResolvers: () => any
	getClientIdsOfDescendants: () => any
	getClientIdsWithDescendants: () => any
	getDraggedBlockClientIds: () => any
	getFirstMultiSelectedBlockClientId: () => any
	getGlobalBlockCount: () => any
	getInserterItems: () => any
	getIsResolving: () => any
	getLastMultiSelectedBlockClientId: () => any
	getLowestCommonAncestorWithSelectedBlock: () => any
	getMultiSelectedBlockClientIds: () => any
	getMultiSelectedBlocks: () => any
	getMultiSelectedBlocksEndClientId: () => any
	getMultiSelectedBlocksStartClientId: () => any
	getNextBlockClientId: ( wrapperClientId: string ) => any
	getPreviousBlockClientId: () => any
	getResolutionError: () => any
	getResolutionState: () => any
	getSelectedBlock: () => any
	getSelectedBlockClientId: () => any
	getSelectedBlockClientIds: () => any
	getSelectedBlockCount: () => any
	getSelectedBlocksInitialCaretPosition: () => any
	getSelectionEnd: () => any
	getSelectionStart: () => any
	getSettings: () => any
	getTemplate: () => any
	getTemplateLock: () => any
	hasBlockMovingClientId: () => any
	hasFinishedResolution: () => any
	hasInserterItems: () => any
	hasMultiSelection: () => any
	hasResolutionFailed: () => any
	hasSelectedBlock: () => any
	hasSelectedInnerBlock: () => any
	hasStartedResolution: () => any
	isAncestorBeingDragged: () => any
	isAncestorMultiSelected: () => any
	isBlockBeingDragged: () => any
	isBlockHighlighted: () => any
	isBlockInsertionPointVisible: () => any
	isBlockMultiSelected: () => any
	isBlockSelected: () => any
	isBlockValid: () => any
	isBlockVisible: () => any
	isBlockWithinSelection: () => any
	isCaretWithinFormattedText: () => any
	isDraggingBlocks: () => any
	isFirstMultiSelectedBlock: () => any
	isLastBlockChangePersistent: () => any
	isMultiSelecting: () => any
	isNavigationMode: () => any
	isResolving: () => any
	isSelectionEnabled: () => any
	isTyping: () => any
	isValidTemplate: () => any
	wasBlockJustInserted: () => any
	__experimentalGetActiveBlockIdByBlockNames: () => any
	__experimentalGetAllowedBlocks: () => any
	__experimentalGetAllowedPatterns: () => any
	__experimentalGetBlockListSettingsForBlocks: () => any
	__experimentalGetDirectInsertBlock: () => any
	__experimentalGetGlobalBlocksByName: () => any
	__experimentalGetLastBlockAttributeChanges: () => any
	__experimentalGetParsedPattern: () => any
	__experimentalGetPatternTransformItems: () => any
	__experimentalGetPatternsByBlockTypes: () => any
	__experimentalGetReusableBlockTitle: () => any
	__unstableGetBlockWithoutInnerBlocks: () => any
	__unstableGetClientIdWithClientIdsTree: () => any
	__unstableGetClientIdsTree: () => any
	__unstableGetContentLockingParent: () => any
	__unstableGetEditorMode: () => any
	__unstableGetSelectedBlocksWithPartialSelection: () => any
	__unstableGetTemporarilyEditingAsBlocks: () => any
	__unstableGetVisibleBlocks: () => any
	__unstableHasActiveBlockOverlayActive: () => any
	__unstableIsFullySelected: () => any
	__unstableIsLastBlockChangeIgnored: () => any
	__unstableIsSelectionCollapsed: () => any
	__unstableIsSelectionMergeable: () => any
	__unstableIsWithinBlockOverlay: () => any
	__unstableSelectionHasUnmergeableBlock: () => any
}